Add and export function can_set_bitrate() in favor of can_calc_bittime()
A default bitrate will now be set when the device is registered and the
device has not already done it. Currenty the default bitrate is defined
by CAN_BITRATE_DEFAULT in linux/can/ioctl.h, but we may want to have an
additional module parameter. Corresponding bitrate settings have been
removed from the device drivers.