cc770: driver for the Bosch CC770 and the Intel AN82527
This patch adds a Socket-CAN driver for the Bosch CC770 and the Intel
AN82527 CAN controllers. Intel stopped production of the AN82527 a few
years ago. The CC770 is a compatible CAN controller from Bosch with
“bugs” being fixed and some improvements in the functionality, mainly:

1. RX and TX error counters are readable.
2. Support of silent (listen-only) mode.
3. Message object 15 can receive all types of frames, also RTR and EFF.

Details are available from Bosch’s „CC770_Product_Info_2007-01.pdf”,
which explains in detail the compatibility between the CC770 and the
82527. This driver use the additional functionality 3. on real CC770
devices. Unfortunately, the CC770 does still not store the message
identifier of the received remote transmission request frame and
therefore it's set to 0.

The driver is derived from the old Socket-CAN i82527 driver in
"drivers/net/can/old/i82527".

Note: This driver reveals a limitation on how to configure the frame
type received by the message object 15. Currently, we do not have an
interface to set controller specific properties. Therefore, the
option can currently be selected globally by the module parameter
"msgobj_eff".

Signed-off-by: Wolfgang Grandegger <wg@grandegger.com>